Utility for boost_static_string
